Use `--fail` flag when downloading the tar file
We don't want to erroneously allow a failed installation to result
in a false-positive, i.e. asdf thinking the operation was successful.

By adding the `--fail` flag, the curl operation will return a non-zero
exit status, resulting in a failure that will invoke the `fail` function.

If we don't do this, even 404s or 500s will dump that error text
into a file with a `.tar` extension that cannot be unpacked.

> tar: Error opening archive: Unrecognized archive format